Migrate to C_Spell.GetSpellInfo for 11.0
The global GetSpellInfo API is deprecated in current clients and removed
in 11.0. fixy fix.

One thing to note is that the new API doesn't accept nil values being
supplied as a parameter - it'll error. This only seems to impact one of
our call sites, but is a fun problem to deal with.