Remove and replace FieldSet template with GroupBox
To deal with the problem of the target bar being multiple colors, the
current fieldset template needs a mixin applied. While we're doing this
we might as well just modernize it entirely (remove names, add parent
keys) and this will break compat for all of its current users, so let's
just rip the bandaid off as quickly as possible here.