Add and make use of new GroupBox template
To deal with the problem of the target bar being multiple colors, the
current fieldset template needs a mixin applied. While we're doing this
we might as well just modernize it entirely (remove names, add parent
keys) and this will break compat for all of its current users, so let's
just rip the bandaid off as quickly as possible here.

The existing fieldset template is retained for compatibility in the
deprecated templates file, but should be cleaned up in Extended over
time.