Adjust fonts without breaking alphabet swaps
The SetFont API is a trap that has the effect of irreversibly breaking
the ability for a fontstring to swap which font file is used for
rendering strings when dealing with text content that goes from one
language to another, eg. English to Russian.

We've had font utilities for a little while now, so let's fix up all the
places calling SetFont and prevent them from breaking.