Change content area background to neutral wood
On some pages the hard edge between our sidebar and the red guild bank
background is quite clear and - frankly - quite ugly. The red is also
used as the background on the configuration pages, which makes text hard
to read in some cases.

The neutral wood asset is in all clients and isn't too crazy a change -
especially as our UI is wood-themed anyway. It's darker and also makes
text easier to read, which is another bonus.