Remove info tooltip size setting
The "Info tooltip size" setting was used to change the size of
informational help tooltips shown in the UI.

This setting has a few issues; firstly, the default size is wrong.
Compared to tooltips around the rest of the game interface, the size is
11 compared to what should be 12 - so by default our tooltips are less
readable for people with bad eyesight.

The more important issue however is that the default value assumes that
font heights across different languages are uniform, which they're not.
Many game fonts in asian languages - particularly Korean - have their
sizes specifically increased to make glyphs clearer.

The adjustment to fonts isn't something that's particularly uniform. At
larger sizes these per-alphabet adjustments eventually stop, but at
smaller sizes they tend to be greater.

As such, we've elected to remove the setting for now and to see who
complains. It should be noted that the game client itself supports DPI
scaling in addition to the UI scaling feature - so users on 4k monitors
and the like already have plenty of knobs they can tune that work across
the entire game, instead of needing to tweak those _and_ our weird
setting.