Fix directory page switching
When swapping between directory profiles, the contents wouldn't change
unless you'd navigated away from the Characteristics tab.

The reason for this is that when showing the page, the selection of the
Characteristics tab is used to refresh the page - however selected tabs
are disabled, so the Click request to select it was being ignored.

Executing the OnClick script directly deals with this problem for now.