Migrate to new menu system (#900)
* Migrate UITools to use new context menus

* Squelch some Luacheck warnings

* Re-add support for legacy context menus

* Add basic compatibility layer for dropdown menus

* Migrate dashboard status panel dropdowns

* Migrate listbox tools to new dropdowns

* Apply dropdown heights OnLoad

* Rework UIDD menu rendering

* Don't register for input by default

* Add Toggle and SetMenuOpen methods

* Add menu response callback support

* Add generic context menu support

* Add UIDD globals to luacheck

* Convert player scan context menu

* Disable dropdowns in configuration again

* Rename AddIconToElement for upstream compat

* Add support for queued descriptions

* Add support for dropdown text overrides

* Propagate mouse motion scripts for DB intrinsics

* Add support for submenu responder overrides

* Port automation menus to new system

* Port text-copy popup to new menu system

* Remove MSA dropdown library

* Be less of an utter donkey

* Remove usages of UIDropDownMenuTemplate

Instead, decorate legacy dropdown buttons manually via region creation
in Lua. This means we eliminate any need to ever give dropdown buttons
names, as UIDropDownMenuTemplate has a mixture of "$parentFoo" names on
its subregions.

Downside is that we now need the dropdown button to be 32 units in
height. The reason for this is that UIDD_Initialize is stupid and sets
the height of the dropdown frame - so unless the size matches, it'd
cause the widget to jump.

* Don't toggle on legacy dropdown clicks

New menus don't toggle, so we shouldn't deviate.

* Port UnitPopup module to new menus

Note that this requires some substantial changes. As part of this,
compatibility with clients that don't have the new menus is entirely
dropped.

This should be a temporary thing ultimately, as eventually the new menu
system will make itself into Classic - but the effort required to
support them both in the context of this module is a bit too much.

* Commit tax fraud

* Shrink hit rect insets for Classic buttons

* Handle undefined UIDD constant in Classic

* Add hack for Classic menu item tooltips

* Visual tweaks to automation dropdowns

* Visual tweaks for legacy dropdowns

* Fix sizing and position of all dropdowns

* Deprecate TRP3_DropDownmenuTemplate

* Toggle menus on click because Classic is dumb

* Use AddInitializer for Classic tooltip fallback

* Make legacy menus play sounds on open and close

* Register for OnEnter and OnLeave by default

This is required by the automation settings which hook the
OnEnter/OnLeave scripts of the dropdown container. There are no
downsides to this otherwise; legacy always has these scripts enabled,
and by default in modern these do nothing.

* Update totalRP3/Deprecated.xml