Misc. tweaks to multiline editboxes
Use a less harsh background for them based upon InsetTemplate3 (which
basically just looks like the old editbox border with a transparent-ish
background), and swap the scrollbars to minimal.

With this scrollbar change, we're now consistently* using the minimal
scrollbar style instead of a mix of minimal + wowtrim.

Additionally fixed the declared height of titled editboxes - the assets
are 20 high, so the frame should also be.