Add utilities for ratio-preserving texture resizes
This adds a few utilities for resizing regions to either fill or fit to
container frames whilst preserving the aspect ratio of the underlying
asset.

The rationale for this is frankly that our UI scales its backgrounds
horrifically; not only are the source assets tiny, but they get
stretched way out of proportion along the horizontal access and look
awful stretched.