Resolve texel snapping issues with new tabs
The left/right anchor pairs which were copied from Blizzards' tab
template exhibit texel snapping issues at certain client UI scales,
resulting in the middle region of the tab sometimes appearing a pixel
above or below the left and right caps.

There's two options to resolve this, the first is to set
texelSnappingBias to 0 and snapToPixelGrid to false. This works fine but
it makes the tabs look blurry at all resolutions as we're downscaling
them.

The other option is to just not use left/right anchors but instead
anchor to the corners, which *should* resolve the issue without blurring
the textures by hopefully guaranteeing that the middle region doesn't
suffer from floating point misprecisions.