Show RP names in dice roll output (#908)
Currently when we receive a dice signal message the character name
printed to the chat is the raw 'sender' string supplied to us by the
comms subsystem, which even if the sender is on the same realm typically
ends up being an ugly as heck "name-realm" string.

Using my superior intellect and absolutely no prodding from someone
else, it has been decided that this is subpar and can be improved. As
such, we now display the senders' RP name and color it according to
their class appropriately.