Currently deserves to be framed. (#909)
* Currently deserves to be framed.

Basically this is based on the "NPC Speech" frame idea.

There's a new toolbar button using the "ui_chat" interface icon, which can be toggled in "Frames settings".

Clicking on it will reveal the Currently frame. Which features a single scrollbox that feeds directly into "CU" (like the profile currently box). Input is sanitzed and version incremented like normal.

There's also a chat command "/trp3 currently" which opens the same chat frame. In case the toolbar is not used or you want to hide the button.

There are three new strings that will require localization.

* Apply suggestions from code review

Who is the real code review villain? Bob or Meorawr?

Co-authored-by: Daniel Yates <me@meorawr.io>

* Apply suggestions from code review

Detective Meorawr is on the case...-ing!

Co-authored-by: Daniel Yates <me@meorawr.io>

* Handle frame toggle so the slash command can benefit from it too

---------

Co-authored-by: Daniel Yates <me@meorawr.io>