Misc. Tooltip Fixups (#920)
* Process tooltip updates correctly again

The previous attempt to fix "ghost" tooltips had the side effect of
breaking their ability to update in response to receiving profile comms,
as we gated the entire check of unit validity behind the primary tooltip
being owned.

This check has been removed; rather than worry about our character
tooltip popping up inappropriately we'll instead just handle it at the
point of anchoring by refusing to anchor the tooltip to the global
GameTooltip unless it itself is owned.

* Add option to disable soft target support

The feature itself is a bit buggy and support is best-effort; we
probably need to take a long hard look at how the tooltip code
fundamentally works before we can support it fully.

While it works *well enough*, let's give people the choice to turn it
off if desired.

* Don't collect original tooltip texts

The original tooltip strings are wholly unused within core at least. The
utility still needs to exist because Extended, but here we can just drop
it.

* Use scripts for clamp inset updates

These work consistently well enough in testing.