Allow reordering of personality traits, now with a button.
The button (and most of its XML) is courtesy of Meorawr who created it, I take no credit.

This has also been implemented (or is it backported?) for Additional Information for consistency. Both are now reordered by the button.

Most of these changes are refactors so that both misc info and traits go through the same functions (like frame.frameIndex and handle.infoTicker being consistent for both).
This includes luals-compatible annotations for the functions I added or refactored.

Two new strings to be localized have been added: `REG_PLAYER_REORDER` and `REG_PLAYER_REORDER_TT`.

NOTE: The cursor will not actually show it is being dragged until you have moved your cursor out of the button you started to drag out of.