Make default RP status OOC
Fixes #179. The default profile is now created for new users with the RP
status field set to OOC, and a flyway has been added to additionally
migrate existing default profiles to OOC.

The rationale for the flyway is just to ensure consistency; it'll be
confusing for all parties involved if we have to explain that for some
users the default state will vary based upon when they first used the
addon, and the only way for users who would want it to default to OOC
would be through manual modification.

The #179 ticket additionally suggests additional tutorials, but it's
unlikely such tutorials would ever be read.