Make RP experience a global setting
Logically - it makes no sense for the RP experience field to be
something that's profile-specific. Beginner users who want to elect into
displaying this status on their profiles need to remember to set it up
manually on each individual one, as it always defaults to experienced.

This changeset makes it such that the RP experience field is now stored
as a configuration variable (not dissimilar to default profile IDs).

All logic for accessing and mutating the XP field has been relocated to the
Player class accordingly. Changing the XP field still results in the
"character" data version incrementing and register-related events
firing, as this is still required for transmission reasons.

As part of transmission, the "character" data is now constructed a bit
more manually and includes the XP field explicitly. The XP field on
player-owned profiles is set to nil as part of a flyway, so there's only
one source of truth for the field.

One downside to this is that the XP field as part of this change will
now effectively reset for all users. It'd be a bit problematic to guess
which experience level is correct if there's a mixture of profiles, so
it's probably best to not do so and instead just make a note in the
changelog.