Update character versions in all profiles
This is required to deal with some edge cases; if a player changes their
experience level and then swaps profiles, other players won't know that
the XP field in the swapped-to profile field is now outdated.

As this field is unlikely to change frequently, we'll just eat the cost
and iterate over all the profiles to update vernums.