Resize about pages on template render
This resolves issues - particularly with T2 - when the user has resized
the window, reloaded, and then views a profile. Without the resize logic
being triggered at the point of rendering, we'd cap the size of the
profile to the constraints of the *original* window size rather than the
larger size the user is actually using.