Add icon templates and put borders on many icons
Icons rendered in profile pages and a few other locations across the UI
are now slightly edge cropped and bordered. This looks rather pretty.

A pair of new templates have been added to support this, both with a
simple SetIconTexture method mixed in that simply accepts an icon name
or ID. If the icon is invalid (per LRPM's view of the world), it will be
sanitized automatically and rendered as the usual misc questionmark.