Rework experience levels
We'll now offer the four following experience levels:

- Newcomer (replaces Beginner)
- Casual (default)
- Veteran (replaces Experienced)
- Newcomer Guide (replaces Volunteer)

The addition of Casual is just because defaulting to Experienced/Veteran
is a bit weird. The terms for Beginner and Volunteer have been changed
to align with Blizzards' terminology as used with the new player
experience stuff - additionally, we now use the same icons as those
where available.