UI Wars Episode 1: The Phantom Backdrop (#937)
* Improve backdrop styling

This commit finally goes around and fixes up all of our backdrop usages
to look consistent; we no longer have a horrific mix of miscellaneous
backdrop sizes or placements dotted around the place, and additionally
we now use the same "gold" color that the achievement frame itself uses
for its own borders.

* Add and make use of new GroupBox template

To deal with the problem of the target bar being multiple colors, the
current fieldset template needs a mixin applied. While we're doing this
we might as well just modernize it entirely (remove names, add parent
keys) and this will break compat for all of its current users, so let's
just rip the bandaid off as quickly as possible here.

The existing fieldset template is retained for compatibility in the
deprecated templates file, but should be cleaned up in Extended over
time.

* Use modern style tabs

Update our tab style to use modern assets. These are available on all
clients as of recent patches, and frankly look a bit better.

Additionally improve the behaviour of tabs when text is truncated in
verbose locales such that we'll display the truncated text in a tooltip
on hover.

* Fix most of the scrollbars

Migrate to the new ScrollFrameTemplate globally, and finally resolve our
long-standing ugliness where scrollbars bleed outside their actual
frames.

This doesn't yet touch scrollbars for faux-scroll lists, like the
profile manager or directory.

* Use modern slider styles

Sliders are only used in the configuration UI. We seem to have an errant
template hanging around that's unused, so also blew that up.

* Use modern checkbox styles

The modern checkbox atlases are available on all clients with Classic
receiving the new settings panel infrastructure, so we can start using
those.

* Use modern editbox styles

* Misc. tweaks to multiline editboxes

Use a less harsh background for them based upon InsetTemplate3 (which
basically just looks like the old editbox border with a transparent-ish
background), and swap the scrollbars to minimal.

Additionally fixed the declared height of titled editboxes - the assets
are 20 high, so the frame should also be.

* Make every damn button 2px taller

The template that all these buttons derives from has a height of 22, not
20. We've been unnecessarily shrinking our buttons, and I don't like
that.

* Various tweaks to main frame backgrounds

Get rid of the ugly red texture that clashes with so much else in the
UI, instead swapping to a nicer more neutral wood where available in
retail.

* Make configuration less ugly in general

Now that the horrific red background is dead, we can do away with the
inset frame entirely and use the minimal scroll bar. This also gives us
a bit more room for the options to render in.

* Apply texture slicing to main menu buttons

This prevents the corners of the main menu bars from getting distorted
when stretched. Mostly improves the look of top-level menu buttons.

* Resize about pages on template render

This resolves issues - particularly with T2 - when the user has resized
the window, reloaded, and then views a profile. Without the resize logic
being triggered at the point of rendering, we'd cap the size of the
profile to the constraints of the *original* window size rather than the
larger size the user is actually using.

* Add UI art assets

These border assets will be used in the rework for various lists around
the UI. Ideally we'd get away with only having half as many files, but
unfortunately texture slicing gained an annoying quirk in 10.2.7 which
means it scales based on frame size rather than being consistent.

* Add icon templates and put borders on many icons

Icons rendered in profile pages and a few other locations across the UI
are now slightly edge cropped and bordered. This looks rather pretty.

A pair of new templates have been added to support this, both with a
simple SetIconTexture method mixed in that simply accepts an icon name
or ID. If the icon is invalid (per LRPM's view of the world), it will be
sanitized automatically and rendered as the usual misc questionmark.

* Make module status page prettier

Use the new assets to tweak the borders of the list elements a little
bit. The background is plain, but this page should eventually be deleted
anyway so not too concerned frankly.

---------

Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>