Make RP experience a global setting (#927)
* Make RP experience a global setting

Logically - it makes no sense for the RP experience field to be
something that's profile-specific. Beginner users who want to elect into
displaying this status on their profiles need to remember to set it up
manually on each individual one, as it always defaults to experienced.

This changeset makes it such that the RP experience field is now stored
as a configuration variable (not dissimilar to default profile IDs).

All logic for accessing and mutating the XP field has been relocated to the
Player class accordingly. Changing the XP field still results in the
"character" data version incrementing and register-related events
firing, as this is still required for transmission reasons.

As part of transmission, the "character" data is now constructed a bit
more manually and includes the XP field explicitly. The XP field on
player-owned profiles is set to nil as part of a flyway, so there's only
one source of truth for the field.

One downside to this is that the XP field as part of this change will
now effectively reset for all users. It'd be a bit problematic to guess
which experience level is correct if there's a mixture of profiles, so
it's probably best to not do so and instead just make a note in the
changelog.

* Update character versions in all profiles

This is required to deal with some edge cases; if a player changes their
experience level and then swaps profiles, other players won't know that
the XP field in the swapped-to profile field is now outdated.

As this field is unlikely to change frequently, we'll just eat the cost
and iterate over all the profiles to update vernums.

* Send XP field as part of vernum exchange

This additionally relocates its storage to be part of the character
table, rather than the profile. The effect of this is that we no longer
need to worry about invalidating all profiles a user owns, and the field
effectively gets treated similarly to the trial character flag.

* Rework experience levels

We'll now offer the four following experience levels:

- Newcomer (replaces Beginner)
- Casual (default)
- Veteran (replaces Experienced)
- Newcomer Guide (replaces Volunteer)

The addition of Casual is just because defaulting to Experienced/Veteran
is a bit weird. The terms for Beginner and Volunteer have been changed
to align with Blizzards' terminology as used with the new player
experience stuff - additionally, we now use the same icons as those
where available.

* Flex that vocab

* Your wish is my command

* On second thought

* Add roleplay experience to misc. profile page

This also takes the opportunity to convert that section of the page to
using parentkeys rather than hard frame names.

* you're breaking my heart rae

* Ensure RP proficiency is consistently hidden