Change unread about icon in tooltips
The note icon that we've been using is a bit confusing. Let's go for a
book-with-an-exclamation-mark, because that's obviously a lot clearer
for users. This doesn't impact the register list (yet).

Additionally, the size of icons in the bottom row of the tooltip have
been increased to make this one asset look good.