Misc roleplay styles revamp + tooltips updated consistency
A lot of information how this decision came to be should be checked in its PR post as it would be too much to note in the full commit.

A summary of these changes would be:
- Removed "In-character frequence" and "Roleplay battle resolution" roleplay styles.
- Added "Accept criminal activity" and "Accept loss of control" roleplay styles.
- Added tooltips on the roleplay styles options (on both the player's own and other profiles).
- Tooltips localizations have been updated to be more consistent, this includes for the tutorial and normal ones (white text + green highlights/keywords).

There is potentially stuff I forgot to put on this commit message, but please do read the PR in full for all information (with images and videos attached).