Improve import and export of translation strings
Following on from the lessons previously learned in #946 and #948, we'll
set the delete-missing-phrases flag _only_ on enUS uploads going forward
as that appears to work fine. All other translation uploads will not set
the flag.

This means that going forward the minimum required process to delete a
translation string should just be to remove it from the 'enUS.lua' file;
a subsequent import of translation strings would then remove it from all
other files automatically. It is, however, totally permissible to remove
the string from all files if that would make things clearer.

Additionally I've made some of the terminology around the place a bit
more consistent for this whole process, eg. preferring to use
"translations" instead of "locales" for target names and workflows. For
the possibility of using non-CF sources (#949) I've also removed
"CurseForge" from the workflow strings.

Finally - one functional change; we no longer import translations on commits
to main. We continue to export new strings from the enUS file, but
that's all. The importing of non-enUS translations continues to happen
on the same timer it always has, as well as being manually runnable.