Misc. UI Tweaks (#950)
* Reset profile page scrollbars on page switch

To try and appease the goddess Rae, we only reset the position when
actually changing the viewed profile - so swapping between tabs or
edit/consult mode won't reset scrollbars, but changing between profiles
A and B will.

* Borderify T2 icons

Would like to do this to T3 as well, but that uses escape sequences for
its textures so more work.

* Increase size of icons in browser

Keep meaning to do this a bit; generally agree with the assessment that
we made icons way too small in the browser. Let's reduce the grid stride
by one and increase the size, and also make them larger in the tooltip.

* Professional padding

The About and More Modules tabs suffer from a bit of "my content hugs
the bottom of my scroll area too tightly". There's probably a few other
areas too.