Misc roleplay styles revamp + tooltips updated consistency (#953)
* Misc roleplay styles revamp + tooltips updated consistency

A lot of information how this decision came to be should be checked in its PR post as it would be too much to note in the full commit.

A summary of these changes would be:
- Removed "In-character frequence" and "Roleplay battle resolution" roleplay styles.
- Added "Accept criminal activity" and "Accept loss of control" roleplay styles.
- Added tooltips on the roleplay styles options (on both the player's own and other profiles).
- Tooltips localizations have been updated to be more consistent, this includes for the tutorial and normal ones (white text + green highlights/keywords).

There is potentially stuff I forgot to put on this commit message, but please do read the PR in full for all information (with images and videos attached).

* Localization type + tooltipframe change.

This was just contained to enUS.

Set tooltipFrame to readOnlyValue so it only triggers when you hover over the user's roleplay style selection text (to be consistent with it only being on dropdown hover for the player).

* Apply suggestions from code review

Who reviewed it better? Solanya or Meorawr? Find out in next week's "reviewing with the lesser ancients"!

Co-authored-by: Daniel Yates <me@meorawr.io>
Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>

---------

Co-authored-by: Daniel Yates <me@meorawr.io>
Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>