Add icon right-click options (copy icon, icon name, paste icon)
Basically adds a right-click dropdown menu to all the locations where the icon browser is available (see the PR for them).
Note: These options work across your player profile and companions.

There are three options:
- Copy icon
Copies the icon from the current frame to an internal clipboard of sorts to be pasted later.
- Copy icon name
Opens a popup where you can quickly copy the icon name to send it to someone or put it in the icon browser if necessary.
- Paste icon
Pastes the icon you had copied earlier with "Copy icon" to the current icon frame.

Four localizations have been added (one does not require new strings, as it is a copy of another that is removed).
Two localizatiosn have been removed.