Update target frame tooltips and companion icons
This project actually expanded a bit in scope when I was working on it.

Initially it was just intended to change the profile buttons (character, companion, mount) but eventually encompassed a lot more. The main result is mostly to make things more consistent on the target frame.

There is 1 new localization, 4 changed and 1 removed.

More information (with images) can be found on the corresponding PR that this will be tied to.