Export RP style options via MSP (#944)
* Export RP style options via MSP

This is to enable their support in MSP-based addons down the line.
Fields are as follows:

- In-character frequency ("RF")
- Accept character injuries ("RI")
- Accept character death ("RD")
- Accept character romance ("RR")
- Roleplay battle resolution ("RB")
- Guild membership ("RG")

Values exported are the same as our dropdown options as-is. May adjust
this based upon feedback, especially as the "RB" option for "TRP roll"
is intrinsically incompatible with other addons.

* Remove contentious fields from export

We're likely removing the IC frequency and battle resolution fields from
recent discussions, so let's not bother with these. For guild membership
status, this field we're not happy with but also don't yet have
solutions for - so let's omit it for now.

* Add criminimal activity and control fields

* Be less stupid