Own companion target frame now has a NPC speech button
This automatically opens the NPC speech frame with the companion's name filled in within [brackets], this only works if you own the companion AND it has a profile attached.

Another addition is auto-focusing on Currently and NPC speech (name or text if name is already filled in) frames.

There are two new LOCs added:
- ``REG_COMPANION_TF_PROFILE_SPEECH
- `REG_COMPANION_TF_PROFILE_SPEECH_TT`