Add walkup button to profiles
I hate XML.

This sort of deals with the issue of where people might not necessarily understand what the icon means on the tooltip.

We present it with a hover tooltip on people's profile so they can hover over it and get used to it consciously (or unconsciously) while it hopefully gets added to people's memory that it means "this person doesn't mind walk ups!".

Also I hate XML, did I say that yet?

This adds one new translation:
- `REG_PLAYER_RESIDENCE_WALKUP_TT`