Add walk-up support and tweak indicator icons (#951)
* Change unread about icon in tooltips

The note icon that we've been using is a bit confusing. Let's go for a
book-with-an-exclamation-mark, because that's obviously a lot clearer
for users. This doesn't impact the register list (yet).

Additionally, the size of icons in the bottom row of the tooltip have
been increased to make this one asset look good.

* Add indicator icon for 'they have notes, dumbo'

Mistress won't let me eat or sleep until this is implemented.

* Use quest ! for unread instead of custom

* Adjust transparency (for spinner) size a bit

* First pass on profile-wide "walkup friendly" icon and setting

* Push the MSP parts so it actually sends/receives

* Add walkup button to profiles

I hate XML.

This sort of deals with the issue of where people might not necessarily understand what the icon means on the tooltip.

We present it with a hover tooltip on people's profile so they can hover over it and get used to it consciously (or unconsciously) while it hopefully gets added to people's memory that it means "this person doesn't mind walk ups!".

Also I hate XML, did I say that yet?

This adds one new translation:
- `REG_PLAYER_RESIDENCE_WALKUP_TT`

* Walkup is Yes or "Do not Show"

* Replaced icons in the directory list

* Here you go Bobby

* Space is the new frontier

* Category title adjustment

* Remove errant space

* for my next trick, I'll make these functions disappear

---------

Co-authored-by: Raenore <172234435+Raenore@users.noreply.github.com>
Co-authored-by: Solanya <solanya@totalrp3.info>