Add additional info + Personality traits duplicate button
This ports over the "Remove" button that both of these options had into a dropdown button.

This dropdown has two options:
- Duplicate
- Delete

Duplicate will make a copy of the current additional info or personality trait line (including custom colors if set), adding it at the bottom of the list.
Delete has been moved lower, and made red, as it is a destructive option. This will hopefully make it much harder to click by accident.

Finally the dropdown button for additional information has been moved a bit to the right so it is on the same horizontal location as the personality trait dropdown.