Refactor menus/dropdowns and duplicate button (#988)
* Refactored dropdowns + More options dropdowns

This is quite big in scope, but it has the following uses:
- Port over all dropdowns using `displayDropDown()` to the new `TRP3_MenuUtil`.
- Create some consistency for the "More options" dropdowns.
- To mimic Blizzard, we use onMouseDown instead of onClick to access these dropdowns.
- Toolbar and TargetFrame buttons can now have an `onMouseDown` property.

This fixes a few things:
- Additive/Creation lines in dropdowns are colored in green.
- Destructive/Remove lines in dropdowns are colored in red.
- Right-click on RP status toolbar (IC/OOC) now has the profile list in alphabetical order.

Other things:
- The icon copy dropdown buttons have been simplified a bit (thanks Meorawr).

Important note:
I am not entirely happy about `onGlanceSlotClick` dropdown rewrite due to how `getGlanceMenuEntries` is set up.
It is obviously still written for how the old dropdown table would receive its info. However currently all the `displayDropDown()` has been removed with all the output being the same or improved where possible.
Refactoring further underlying code should happen at some point, so that snippets like that don't have to exist.

Three new localization keys:
- `CM_OPTIONS`
- `CM_OPTIONS_ADDITIONAL`
- `CM_DUPLICATE`

* Add additional info + Personality traits duplicate button

This ports over the "Remove" button that both of these options had into a dropdown button.

This dropdown has two options:
- Duplicate
- Delete

Duplicate will make a copy of the current additional info or personality trait line (including custom colors if set), adding it at the bottom of the list.
Delete has been moved lower, and made red, as it is a destructive option. This will hopefully make it much harder to click by accident.

Finally the dropdown button for additional information has been moved a bit to the right so it is on the same horizontal location as the personality trait dropdown.

* Fix module options button

* Apply suggestions from code review

To capitalize or not capitalize.. That is THE question!

Co-authored-by: Daniel Yates <me@meorawr.io>

* Apply suggestions from code review

Hocus pocus bugs are gone and you all saw nothing. (Thank you, Sol.)

Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>

* Appease the Bobster

---------

Co-authored-by: Daniel Yates <me@meorawr.io>
Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>
Co-authored-by: Solanya <solanya@totalrp3.info>