Convert preset additional info & traits to custom
Basically allows you to convert a preset additional info or trait to be custom.
On conversion, it will keep its icon, values, names, etc. It just becomes customizeable.

Note: The action button SetScript had to be moved out of the `if frame == nil then` block as it would otherwise not know its IDs got changed, which in turn would make it never know to stop hiding "convert to custom" from the user.

If an additional info was one that added something to the tooltip, the user gets a warning that mentions it will be gone after conversion.
For the above behavior to work, a `shownOnTooltip` boolean field was added to all `MiscInfoTypeData`.

Two new localizations have been added:
- `REG_PLAYER_CONVERT`
- `REG_PLAYER_CONVERT_WARNING`