Fix positioning of regions on music player frame (#1004)
The "URL" text in particular (which is really just the music name
string) could sometimes position itself awkwardly and end up behind the
buttons that were beneath it. This should be resolved by disabling
wrapping and forcing the text to vertically align to the top.

Additionally slightly tweaked the play and stop buttons to be more
consistently placed.