Convert profile managers to scrollbox list views
This converts the player and companion profile managers to use scrollbox
list views rather than the old and crappy UI list system that forced
those things to be memes for years and only support a fixed number of
frames.

The code is a mess, but that's because frankly rewriting the entire
profile manager would take a lot longer than we really have time for.
Additionally while there's some changes for the profile importer, that
has not been converted to a list view because it's another special beast
entirely. I've no idea how that tab even looks right now with these
changes.