Don't show minimum frame size as invalid (#1007)
The resize frame logic doesn't too consistently round its height and
width values when updating the size text. Somehow this results in it
treating the minimum size (840x600) as invalid and renders it in red.

Rounding values up a bit earlier for some reason resolves this. Not sure
why, don't care why.