Fix issues with enabling/disabling modules
Truly module management is something that can never be truly "fixed",
but we can at least maintain a basic facade about it being basically
functional from time to time.

The dropdown reworks haven't been kind to this code and it simply
assumes that a path from the responder to the module list item exists
like it used to.